Configuration revamp
Create a proper generic API of configuration, completely independent
than other layers of the plugin. Allowing a layered structure of the
different subsystems in the code.

The basic ideas is as follows:

The configuration API provides a couple of function (excluding the data
directory creation function): one for getting a value and another for
setting a value. All configuration clients should use this functionality
to handle their configuration settings.This approach resolves all issues
related to configuration access, as the `get` and `set` functions handle
reading from and writing to the configuration file as needed. Clients do
not need to be aware of the file's existence or structure.

For example, the model selection functionality is now a client of the
new configuration API. Instead of interacting directly with the
configuration, it uses the API to retrieve and update configuration
values (See subsequent commits for details on the separation of models
and configuration).

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>