Global variables not in autoload
Global variables should all be declared in the main plugin file, and not
in any of the auto loads, as code might refer to variable that is
declared in an unloaded autoload file, leading to errors. (There's
already a weird close/wipeout sequence that does it).

Better to move things now to prevent future bugs and have these
variables placed and declared properly.

Also uses the 'get' over 'g:' functionality instead of the 'exists'
idiom. Makes clear and shorter code.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>